+++ GGDeals/Api/Models/GameWithLauncher.cs
@@ -1,0 +1,9 @@
+﻿using Playnite.SDK.Models;
+
+namespace GGDeals.Api.Models
+{
+	public class GameWithLauncher : Game
+	{
+		public GGLauncher GGLauncher { get; set; }
+	}
+}
