+++ GGDeals/Api/Models/ImportResponse.cs
@@ -1,0 +1,9 @@
+﻿namespace GGDeals.Api.Models
+{
+	public class ImportResponse
+	{
+		public bool Success { get; set; }
+
+		public ResponseData Data { get; set; }
+	}
+}
