+++ GGDeals/Api/Models/ImportResultStatus.cs
@@ -1,0 +1,10 @@
+﻿namespace GGDeals.Api.Models
+{
+	public enum ImportResultStatus
+	{
+		Error,
+		Added,
+		Skipped,
+		Miss,
+	}
+}
