+++ GGDeals/Api/Models/ResponseData.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace GGDeals.Api.Models
+{
+	public class ResponseData
+	{
+		public List<ImportResult> Result { get; set; }
+	}
+}
